+++ src/app/virtualmachines/clients/client.model.ts
@@ -16,7 +16,12 @@ export class Client {
     private _maxVolumeLimit: number;
     private _assignedVolumesStorage: number;
 
-    constructor() {
+
+    constructor(host: string, port: string, location: string, id: string) {
+        this._host = host;
+        this._port = port;
+        this._location = location;
+        this._id = id;
     }
 
     get id(): string {
