+++ src/app/virtualmachines/clients/clientOverview.component.ts
@@ -1,6 +1,6 @@
 import {Component, OnInit} from '@angular/core';
-import {Client} from './vmclient';
-import {ClientService} from '../../api-connector/vmClients.service';
+import {Client} from './client.model';
+import {ClientService} from '../../api-connector/client.service';
 import {ApiSettings} from '../../api-connector/api-settings.service';
 import {GroupService} from '../../api-connector/group.service';
 import {UserService} from '../../api-connector/user.service';
@@ -13,7 +13,7 @@ import {VoService} from '../../api-connector/vo.service';
  */
 @Component({
     selector: 'app-client-overview',
-    templateUrl: 'vmClients.component.html',
+    templateUrl: 'clientOverview.html',
     providers: [FacilityService, VoService, UserService, GroupService, ClientService, ApiSettings]
 })
 
@@ -141,6 +141,7 @@ export class ClientOverviewComponent implements OnInit {
         this.checkVOstatus();
         this.getClientsChecked();
         this.getComputeCenters();
+        this.clientservice.updateClient(new Client('test', '8432', 'da', '2')).subscribe()
 
     }
