+++ src/app/virtualmachines/vm.module.ts
@@ -8,7 +8,7 @@ import {ModalModule} from 'ngx-bootstrap/modal';
 import {ImageDetailComponent} from '../virtualmachines/imagedetail.component';
 import {VirtualMachineComponent} from '../virtualmachines/addvm.component';
 import {FlavorDetailComponent} from '../virtualmachines/flavordetail.component';
-import {ClientOverviewComponent} from './clients/vmClients.component';
+import {ClientOverviewComponent} from './clients/clientOverview.component';
 import {VmOverviewComponent} from '../virtualmachines/vmOverview.component';
 import {VolumeOverviewComponent} from './volumes/volumeOverview.component';
 import {SnapshotOverviewComponent} from './snapshots/snapshotOverview.component';
