+++ src/app/virtualmachines/vm_routing.module.ts
@@ -1,7 +1,7 @@
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
 import {VirtualMachineComponent} from '../virtualmachines/addvm.component';
-import {ClientOverviewComponent} from './clients/vmClients.component';
+import {ClientOverviewComponent} from './clients/clientOverview.component';
 import {VmOverviewComponent} from '../virtualmachines/vmOverview.component';
 import {VolumeOverviewComponent} from './volumes/volumeOverview.component';
 import {SnapshotOverviewComponent} from './snapshots/snapshotOverview.component';
